+++ apps/prairielearn/elements/pl-code/code_utils.py
@@ -1,15 +1,12 @@
-from typing import Optional
-
-
-def parse_highlight_lines(highlight_lines: str) -> Optional[list[int]]:
+def parse_highlight_lines(highlight_lines: str) -> list[int] | None:
     """
     Parses a string like "1", "1-4", "1-3,5,7-8" into a list of lines like
     [1], [1,2,3,4], and [1,2,3,5,7,8]
     """
     lines = []
     components = highlight_lines.split(",")
-    for component in components:
-        component = component.replace(" ", "").split("-")
+    for raw_component in components:
+        component = raw_component.replace(" ", "").split("-")
         try:
             if len(component) == 1:
                 line = int(component[0])