+++ apps/prairielearn/src/lib/code-caller/index.ts
@@ -1,4 +1,3 @@
-// @ts-check
 import * as os from 'node:os';
 
 import debugfn from 'debug';
@@ -53,22 +52,27 @@ export async function init() {
     {
       create: async () => {
         const codeCallerOptions = {
+          dropPrivileges: false,
           questionTimeoutMilliseconds: config.questionTimeoutMilliseconds,
           pingTimeoutMilliseconds: config.workerPingTimeoutMilliseconds,
-          errorLogger: logger.error.bind(logger),
         };
 
-        const codeCaller = run(() => {
+        const codeCaller = await run(async () => {
           if (workersExecutionMode === 'container') {
-            return new CodeCallerContainer(codeCallerOptions);
+            return await CodeCallerContainer.create(codeCallerOptions);
           } else if (workersExecutionMode === 'native') {
-            return new CodeCallerNative(codeCallerOptions);
+            return await CodeCallerNative.create({
+              ...codeCallerOptions,
+              pythonVenvSearchPaths: config.pythonVenvSearchPaths,
+              errorLogger: logger.error.bind(logger),
+              // We can only drop privileges if this code caller is running in a container.
+              dropPrivileges: false,
+            });
           } else {
             throw new Error(`Unexpected workersExecutionMode: ${workersExecutionMode}`);
           }
         });
 
-        await codeCaller.ensureChild();
         load.startJob('python_worker_idle', codeCaller.uuid);
         return codeCaller;
       },
@@ -223,4 +227,4 @@ export async function withCodeCaller<T>(
   }
 }
 
-export { FunctionMissingError, CodeCaller };
+export { FunctionMissingError, type CodeCaller };