+++ apps/prairielearn/src/tests/permissions/courseAdminAccess.test.ts
@@ -124,15 +124,14 @@ function runTest(context) {
       response.$,
       'button[data-testid="add-users-button"]',
     );
-    const form = {
-      __action: 'course_permissions_insert_by_user_uids',
-      __csrf_token: context.__csrf_token,
-      uid: ' staff03@example.com ,   ,   staff04@example.com',
-      course_role: 'Viewer',
-    };
     response = await helperClient.fetchCheerio(context.pageUrl, {
       method: 'POST',
-      form,
+      body: new URLSearchParams({
+        __action: 'course_permissions_insert_by_user_uids',
+        __csrf_token: context.__csrf_token,
+        uid: ' staff03@example.com ,   ,   staff04@example.com',
+        course_role: 'Viewer',
+      }),
       headers,
     });
     assert.isTrue(response.ok);
@@ -151,15 +150,14 @@ function runTest(context) {
       response.$,
       'button[data-testid="add-users-button"]',
     );
-    const form = {
-      __action: 'course_permissions_insert_by_user_uids',
-      __csrf_token: context.__csrf_token,
-      uid: `staff03@example.com, staff05@example.com, ${new_user}`,
-      course_role: 'None',
-    };
     response = await helperClient.fetchCheerio(context.pageUrl, {
       method: 'POST',
-      form,
+      body: new URLSearchParams({
+        __action: 'course_permissions_insert_by_user_uids',
+        __csrf_token: context.__csrf_token,
+        uid: `staff03@example.com, staff05@example.com, ${new_user}`,
+        course_role: 'None',
+      }),
       headers,
     });
     assert.isTrue(response.ok);
@@ -178,15 +176,14 @@ function runTest(context) {
       response.$,
       'form[name=student-data-access-add-3]',
     );
-    const form = {
-      __action: 'course_instance_permissions_insert',
-      __csrf_token: context.__csrf_token,
-      user_id: 3,
-      course_instance_id: 1,
-    };
     response = await helperClient.fetchCheerio(context.pageUrl, {
       method: 'POST',
-      form,
+      body: new URLSearchParams({
+        __action: 'course_instance_permissions_insert',
+        __csrf_token: context.__csrf_token,
+        user_id: '3',
+        course_instance_id: '1',
+      }),
       headers,
     });
     assert.isTrue(response.ok);
@@ -204,14 +201,13 @@ function runTest(context) {
       response.$,
       'form[name=course-content-access-form-3]',
     );
-    const form = {
-      __action: 'course_permissions_delete',
-      __csrf_token: context.__csrf_token,
-      user_id: 3,
-    };
     response = await helperClient.fetchCheerio(context.pageUrl, {
       method: 'POST',
-      form,
+      body: new URLSearchParams({
+        __action: 'course_permissions_delete',
+        __csrf_token: context.__csrf_token,
+        user_id: '3',
+      }),
       headers,
     });
     assert.isTrue(response.ok);
@@ -226,14 +222,13 @@ function runTest(context) {
     assert.isTrue(response.ok);
     const __csrf_token = response.$('span[id=test_csrf_token]').text();
     assert.lengthOf(response.$(`form[name=course-content-access-form-${context.userId}]`), 0);
-    const form = {
-      __action: 'course_permissions_delete',
-      __csrf_token,
-      user_id: 2,
-    };
     response = await helperClient.fetchCheerio(context.pageUrl, {
       method: 'POST',
-      form,
+      body: new URLSearchParams({
+        __action: 'course_permissions_delete',
+        __csrf_token,
+        user_id: '2',
+      }),
       headers,
     });
     assert.equal(response.status, 403);
@@ -250,15 +245,14 @@ function runTest(context) {
       response.$,
       'form[name=course-content-access-form-4]',
     );
-    const form = {
-      __action: 'course_permissions_update_role',
-      __csrf_token: context.__csrf_token,
-      user_id: 4,
-      course_role: 'Owner',
-    };
     response = await helperClient.fetchCheerio(context.pageUrl, {
       method: 'POST',
-      form,
+      body: new URLSearchParams({
+        __action: 'course_permissions_update_role',
+        __csrf_token: context.__csrf_token,
+        user_id: '4',
+        course_role: 'Owner',
+      }),
       headers,
     });
     assert.isTrue(response.ok);
@@ -273,15 +267,14 @@ function runTest(context) {
     assert.isTrue(response.ok);
     const __csrf_token = response.$('span[id=test_csrf_token]').text();
     assert.lengthOf(response.$(`form[name=course-content-access-form-${context.userId}]`), 0);
-    const form = {
-      __action: 'course_permissions_update_role',
-      __csrf_token,
-      user_id: context.userId,
-      course_role: 'None',
-    };
     response = await helperClient.fetchCheerio(context.pageUrl, {
       method: 'POST',
-      form,
+      body: new URLSearchParams({
+        __action: 'course_permissions_update_role',
+        __csrf_token,
+        user_id: context.userId,
+        course_role: 'None',
+      }),
       headers,
     });
     assert.equal(response.status, 403);
@@ -298,14 +291,13 @@ function runTest(context) {
     assert.isTrue(response.ok);
     const __csrf_token = response.$('span[id=test_csrf_token]').text();
     assert.lengthOf(response.$(`form[name=course-content-access-form-${context.userId}]`), 0);
-    const form = {
-      __action: 'course_permissions_delete',
-      __csrf_token,
-      user_id: context.userId,
-    };
     response = await helperClient.fetchCheerio(context.pageUrl, {
       method: 'POST',
-      form,
+      body: new URLSearchParams({
+        __action: 'course_permissions_delete',
+        __csrf_token,
+        user_id: context.userId,
+      }),
       headers,
     });
     assert.equal(response.status, 403);
@@ -322,15 +314,14 @@ function runTest(context) {
     assert.isTrue(response.ok);
     const __csrf_token = response.$('span[id=test_csrf_token]').text();
     assert.lengthOf(response.$(`form[name=course-content-access-form-${context.userId}]`), 0);
-    const form = {
-      __action: 'course_permissions_update_role',
-      __csrf_token,
-      user_id: context.userId,
-      course_role: 'None',
-    };
     response = await helperClient.fetchCheerio(context.pageUrl, {
       method: 'POST',
-      form,
+      body: new URLSearchParams({
+        __action: 'course_permissions_update_role',
+        __csrf_token,
+        user_id: context.userId,
+        course_role: 'None',
+      }),
       headers,
     });
     assert.equal(response.status, 403);
@@ -347,15 +338,14 @@ function runTest(context) {
       response.$,
       'button[data-testid="add-users-button"]',
     );
-    const form = {
-      __action: 'course_permissions_insert_by_user_uids',
-      __csrf_token: context.__csrf_token,
-      uid: 'staff03@example.com',
-      course_role: 'None',
-    };
     response = await helperClient.fetchCheerio(context.pageUrl, {
       method: 'POST',
-      form,
+      body: new URLSearchParams({
+        __action: 'course_permissions_insert_by_user_uids',
+        __csrf_token: context.__csrf_token,
+        uid: 'staff03@example.com',
+        course_role: 'None',
+      }),
       headers,
     });
     assert.isTrue(response.ok);
@@ -373,15 +363,14 @@ function runTest(context) {
       response.$,
       'form[name=student-data-access-add-3]',
     );
-    const form = {
-      __action: 'course_instance_permissions_insert',
-      __csrf_token: context.__csrf_token,
-      user_id: 3,
-      course_instance_id: 1,
-    };
     response = await helperClient.fetchCheerio(context.pageUrl, {
       method: 'POST',
-      form,
+      body: new URLSearchParams({
+        __action: 'course_instance_permissions_insert',
+        __csrf_token: context.__csrf_token,
+        user_id: '3',
+        course_instance_id: '1',
+      }),
       headers,
     });
     assert.isTrue(response.ok);
@@ -399,16 +388,15 @@ function runTest(context) {
       response.$,
       'form[name=student-data-access-change-3-1]',
     );
-    const form = {
-      __action: 'course_instance_permissions_update_role_or_delete',
-      __csrf_token: context.__csrf_token,
-      user_id: 3,
-      course_instance_id: 1,
-      course_instance_role: 'Student Data Editor',
-    };
     response = await helperClient.fetchCheerio(context.pageUrl, {
       method: 'POST',
-      form,
+      body: new URLSearchParams({
+        __action: 'course_instance_permissions_update_role_or_delete',
+        __csrf_token: context.__csrf_token,
+        user_id: '3',
+        course_instance_id: '1',
+        course_instance_role: 'Student Data Editor',
+      }),
       headers,
     });
     assert.isTrue(response.ok);
@@ -426,15 +414,14 @@ function runTest(context) {
       response.$,
       'form[name=student-data-access-add-5]',
     );
-    const form = {
-      __action: 'course_instance_permissions_insert',
-      __csrf_token: context.__csrf_token,
-      user_id: 5,
-      course_instance_id: 1,
-    };
     response = await helperClient.fetchCheerio(context.pageUrl, {
       method: 'POST',
-      form,
+      body: new URLSearchParams({
+        __action: 'course_instance_permissions_insert',
+        __csrf_token: context.__csrf_token,
+        user_id: '5',
+        course_instance_id: '1',
+      }),
       headers,
     });
     assert.isTrue(response.ok);
@@ -452,15 +439,14 @@ function runTest(context) {
       response.$,
       'form[name=student-data-access-change-5-1]',
     );
-    const form = {
-      __action: 'course_instance_permissions_update_role_or_delete',
-      __csrf_token: context.__csrf_token,
-      user_id: 5,
-      course_instance_id: 1,
-    };
     response = await helperClient.fetchCheerio(context.pageUrl, {
       method: 'POST',
-      form,
+      body: new URLSearchParams({
+        __action: 'course_instance_permissions_update_role_or_delete',
+        __csrf_token: context.__csrf_token,
+        user_id: '5',
+        course_instance_id: '1',
+      }),
       headers,
     });
     assert.isTrue(response.ok);
@@ -478,13 +464,12 @@ function runTest(context) {
       response.$,
       'button[data-testid="remove-all-student-data-access-button"]',
     );
-    const form = {
-      __action: 'remove_all_student_data_access',
-      __csrf_token: context.__csrf_token,
-    };
     response = await helperClient.fetchCheerio(context.pageUrl, {
       method: 'POST',
-      form,
+      body: new URLSearchParams({
+        __action: 'remove_all_student_data_access',
+        __csrf_token: context.__csrf_token,
+      }),
       headers,
     });
     assert.isTrue(response.ok);
@@ -502,15 +487,14 @@ function runTest(context) {
       response.$,
       'form[name=student-data-access-add-5]',
     );
-    const form = {
-      __action: 'course_instance_permissions_insert',
-      __csrf_token: context.__csrf_token,
-      user_id: 5,
-      course_instance_id: 1,
-    };
     response = await helperClient.fetchCheerio(context.pageUrl, {
       method: 'POST',
-      form,
+      body: new URLSearchParams({
+        __action: 'course_instance_permissions_insert',
+        __csrf_token: context.__csrf_token,
+        user_id: '5',
+        course_instance_id: '1',
+      }),
       headers,
     });
     assert.isTrue(response.ok);
@@ -528,13 +512,12 @@ function runTest(context) {
       response.$,
       'button[data-testid="delete-users-with-no-access-button"]',
     );
-    const form = {
-      __action: 'delete_no_access',
-      __csrf_token: context.__csrf_token,
-    };
     response = await helperClient.fetchCheerio(context.pageUrl, {
       method: 'POST',
-      form,
+      body: new URLSearchParams({
+        __action: 'delete_no_access',
+        __csrf_token: context.__csrf_token,
+      }),
       headers,
     });
     assert.isTrue(response.ok);
@@ -553,13 +536,12 @@ function runTest(context) {
       response.$,
       'button[data-testid="delete-non-owners-button"]',
     );
-    const form = {
-      __action: 'delete_non_owners',
-      __csrf_token: context.__csrf_token,
-    };
     response = await helperClient.fetchCheerio(context.pageUrl, {
       method: 'POST',
-      form,
+      body: new URLSearchParams({
+        __action: 'delete_non_owners',
+        __csrf_token: context.__csrf_token,
+      }),
       headers,
     });
     assert.isTrue(response.ok);
@@ -577,15 +559,14 @@ function runTest(context) {
       response.$,
       'form[name=course-content-access-form-4]',
     );
-    const form = {
-      __action: 'course_permissions_update_role',
-      __csrf_token: context.__csrf_token,
-      user_id: 4,
-      course_role: 'Editor',
-    };
     response = await helperClient.fetchCheerio(context.pageUrl, {
       method: 'POST',
-      form,
+      body: new URLSearchParams({
+        __action: 'course_permissions_update_role',
+        __csrf_token: context.__csrf_token,
+        user_id: '4',
+        course_role: 'Editor',
+      }),
       headers,
     });
     assert.isTrue(response.ok);